+++ packages/common/src/directives/ng_optimized_image/image_loaders/cloudinary_loader.ts
@@ -10,9 +10,9 @@ import {Provider, ɵRuntimeError as RuntimeError} from '@angular/core';
 
 import {RuntimeErrorCode} from '../../../errors';
 import {PRECONNECT_CHECK_BLOCKLIST} from '../preconnect_link_checker';
+import {isValidPath, normalizePath, normalizeSrc} from '../util';
 
 import {IMAGE_LOADER, ImageLoaderConfig} from './image_loader';
-import {isValidPath, normalizePath, normalizeSrc} from './loader_utils';
 
 /**
  * Function that generates a built-in ImageLoader for Cloudinary
