+++ packages/common/src/directives/ng_optimized_image/util.ts
@@ -38,3 +38,26 @@ export function extractHostname(url: string): string {
   }
   return url;
 }
+
+export function isValidPath(path: unknown): boolean {
+  const isString = typeof path === 'string';
+
+  if (!isString || path.trim() === '') {
+    return false;
+  }
+
+  try {
+    const url = new URL(path);
+    return true;
+  } catch {
+    return false;
+  }
+}
+
+export function normalizePath(path: string): string {
+  return path.endsWith('/') ? path.slice(0, -1) : path;
+}
+
+export function normalizeSrc(src: string): string {
+  return src.startsWith('/') ? src.slice(1) : src;
+}
